pyproject.toml
[project]
name = "meta-ads-mcp"
version = "0.1.0"
description = "MCP server for managing Meta Ads campaigns"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "facebook-business>=25.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.2",
]

[project.scripts]
meta-ads-mcp = "meta_ads_mcp.server:main"
validate-meta-campaign = "meta_ads_mcp.campaign_schema:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meta_ads_mcp"]
